refactor: move/clear test code
